requirements.txt
asgiref==3.2.7
cloudinary==1.22.0
cloudpickle==1.3.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.4
Django==3.0.4
django-allauth==0.42.0
django-bootstrap-form==3.4
django-braces==1.14.0
django-ckeditor==6.0.0
django-comments-xtd==2.6.1
django-contrib-comments==1.9.2
django-cors-headers==3.2.1
django-crispy-forms==1.9.2
django-etc==1.2.0
django-filter==2.2.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-ipware==3.0.2
django-js-asset==1.2.2
django-model-utils==4.0.0
django-modeltranslation==0.14.4
django-recaptcha3==0.1.0
django-rest-auth==0.9.5
django-star-ratings==0.9.1
django-storages==1.9.1
django-summernote==0.8.11.6
django-tinymce==3.2.0
django-tinymce4-lite==1.7.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.15.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
gunicorn==20.0.4
jmespath==0.9.5
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2019.3
requests-oauthlib==1.3.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
soupsieve==2.0
sqlparse==0.3.1
statsmodels==0.10.2
swapper==1.1.2.post1
termcolor==1.1.0
w3lib==1.21.0
webdrivermanager==0.9.0
websocket-client==0.57.0
whitenoise==5.2.0